21w43a

+++ net/minecraft/client/render/debug/HeightmapDebugRenderer.java
@@ -59,7 +59,7 @@ implements DebugRenderer.Renderer {
                             int m = ChunkSectionPos.getOffsetPos(chunkPos.x, k);
                             int n = ChunkSectionPos.getOffsetPos(chunkPos.z, l);
                             float f = (float)((double)((float)worldAccess.getTopY(type, m, n) + (float)type.ordinal() * 0.09375f) - cameraY);
-                            WorldRenderer.drawBox(bufferBuilder, (double)((float)m + 0.25f) - cameraX, f, (double)((float)n + 0.25f) - cameraZ, (double)((float)m + 0.75f) - cameraX, f + 0.09375f, (double)((float)n + 0.75f) - cameraZ, vec3f.getX(), vec3f.getY(), vec3f.getZ(), 1.0f);
+                            WorldRenderer.drawBox(bufferBuilder, (double)((float)m + 0.25f) - cameraX, (double)f, (double)((float)n + 0.25f) - cameraZ, (double)((float)m + 0.75f) - cameraX, (double)(f + 0.09375f), (double)((float)n + 0.75f) - cameraZ, vec3f.getX(), vec3f.getY(), vec3f.getZ(), 1.0f);
                         }
                     }
                 }
